Final build
Work was attempted on 3D tessellation of character model, but was unsuccessful - removed all code relevant to this. Changed background room material to a shiny gold, to add a 'regal' vibe to the scene. Fixed texturing issues present in CW1, so character model now displays as intended - as a result, does not have PBR relevant materials such as shininess etc but the intention for this was to make the background room shiny rather than the character model. Finally, also fixed issues with .exe build - had to perform linking in the Release build in VS Solution, and also transfer over project files to Release directory.